Moved all topology code to a new file cluster_topology.rs
